refactor io config to storage options
