+++ daft/daft.pyi
@@ -1420,7 +1420,6 @@ class LogicalPlanBuilder:
         path: str,
         columns_name: list[str],
         mode: str,
-        current_version: int,
         large_dtypes: bool,
         io_config: IOConfig | None = None,
     ) -> LogicalPlanBuilder: ...
