+++ daft/execution/execution_step.py
@@ -412,7 +412,6 @@ class WriteIceberg(SingleOutputInstruction):
 class WriteDeltaLake(SingleOutputInstruction):
     base_path: str
     large_dtypes: bool
-    current_version: int
     io_config: IOConfig | None
 
     def run(self, inputs: list[MicroPartition]) -> list[MicroPartition]:
@@ -439,7 +438,6 @@ class WriteDeltaLake(SingleOutputInstruction):
             input,
             large_dtypes=self.large_dtypes,
             base_path=self.base_path,
-            current_version=self.current_version,
             io_config=self.io_config,
         )
