+++ daft/io/_lance.py
@@ -6,10 +6,10 @@ from daft import context
 from daft.api_annotations import PublicAPI
 from daft.daft import IOConfig, Pushdowns, PyTable, ScanOperatorHandle, ScanTask
 from daft.dataframe import DataFrame
-from daft.io.object_store_options import io_config_to_storage_options
 from daft.io.scan import PartitionField, ScanOperator
 from daft.logical.builder import LogicalPlanBuilder
 from daft.logical.schema import Schema
+from daft.object_store_options import io_config_to_storage_options
 from daft.table import Table
 
 if TYPE_CHECKING:
