+++ src/daft-plan/src/builder.rs
@@ -511,7 +511,6 @@ impl LogicalPlanBuilder {
         path: String,
         columns_name: Vec<String>,
         mode: String,
-        current_version: i32,
         large_dtypes: bool,
         io_config: Option<IOConfig>,
     ) -> DaftResult<Self> {
@@ -520,7 +519,6 @@ impl LogicalPlanBuilder {
             catalog: crate::sink_info::CatalogType::DeltaLake(DeltaLakeCatalogInfo {
                 path,
                 mode,
-                current_version,
                 large_dtypes,
                 io_config,
             }),
@@ -805,7 +803,6 @@ impl PyLogicalPlanBuilder {
         path: String,
         columns_name: Vec<String>,
         mode: String,
-        current_version: i32,
         large_dtypes: bool,
         io_config: Option<common_io_config::python::IOConfig>,
     ) -> PyResult<Self> {
@@ -815,7 +812,6 @@ impl PyLogicalPlanBuilder {
                 path,
                 columns_name,
                 mode,
-                current_version,
                 large_dtypes,
                 io_config.map(|cfg| cfg.config),
             )?
