+++ src/daft-plan/src/physical_plan.rs
@@ -751,7 +751,6 @@ fn deltalake_write(
             upstream_iter,
             &delta_lake_info.path,
             delta_lake_info.large_dtypes,
-            delta_lake_info.current_version,
             delta_lake_info
                 .io_config
                 .as_ref()
