+++ tests/io/delta_lake/conftest.py
@@ -19,7 +19,7 @@ from pytest_lazyfixture import lazy_fixture
 
 import daft
 from daft import DataCatalogTable, DataCatalogType
-from daft.io.object_store_options import io_config_to_storage_options
+from daft.object_store_options import io_config_to_storage_options
 from tests.io.delta_lake.mock_aws_server import start_service, stop_process
